vim: split plugin setup into multiple files
