+++ vim/rtp/lua/plugins/commandline.lua
@@ -1,0 +1,3 @@
+return {
+	"vim-utils/vim-husk",
+}
