+++ vim/rtp/lua/plugins/debug.lua
@@ -1,0 +1,9 @@
+return {
+	"mfussenegger/nvim-dap",
+	{
+		"rcarriga/nvim-dap-ui",
+		dependencies = {
+			"mfussenegger/nvim-dap",
+		},
+	},
+}
