+++ vim/rtp/lua/plugins/quickfix.lua
@@ -1,0 +1,3 @@
+return {
+	"kevinhwang91/nvim-bqf",
+}
